add relation between posts and user + user can only modify what they have created

+++ controllers/feed.js
@@ -2,6 +2,7 @@ const { validationResult } = require("express-validator/check")
 const Post = require('./../models/post');
 const fs = require("fs")
 const path = require("path");
+const User = require("./../models/user");
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1
@@ -28,7 +29,6 @@ exports.getPosts = (req, res, next) => {
             }
             next(err);
         });
-
 }
 
 exports.postPost = (req, res, next) => {
@@ -46,20 +46,31 @@ exports.postPost = (req, res, next) => {
     const title = req.body.title;
     const content = req.body.content;
     const imageUrl = req.file.path.replace("\\", "/");
-
+    let creator;
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: { name: 'Satoshi' }
+        creator: req.userId
     });
     post
         .save()
         .then(result => {
-            console.log(result)
+            return User.findById(req.userId)
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post)
+            return user.save()
+        })
+        .then(result => {
             res.status(201).json({
                 message: 'Post created successfully!',
-                post: result
+                post: post,
+                creator: {
+                    _id: creator._id,
+                    name: creator.name
+                }
             });
         })
         .catch(err => {
@@ -121,6 +132,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw (error)
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error("unauthorized")
+                error.statusCode = 403;
+                throw error
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl)
             }
@@ -153,12 +169,23 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw (error)
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error("unauthorized")
+                error.statusCode = 403;
+                throw error
+            }
             //check logged in user
             clearImage(post.imageUrl)
             return Post.findByIdAndRemove(postId)
         })
         .then(result => {
-            console.log(result)
+            return User.findById(req.userId)
+        })
+        .then(user => {
+            user.posts.pull(postId)
+            return user.save()
+        })
+        .then(result => {
             res.status(200).json({
                 message: "deleted successfully"
             })
